% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADILSONVOL2CAP20EXEMP1.R
\docType{data}
\name{ADILSONVOL2CAP20EXEMP1}
\alias{ADILSONVOL2CAP20EXEMP1}
\title{ADILSONVOL2CAP20EXEMP1}
\format{
A data frame with 11 rows and 2 variables:
\describe{
  \item{LOCAL}{PAIS}
  \item{TAXA DE MORTALIDADE}{TAXA DE MORTALIDADE A CADA UM MILHAO DE HABITANTES}
}
}
\usage{
ADILSONVOL2CAP20EXEMP1
}
\description{
MORTES CAUSADAS POR USO DE DROGAS
}
\keyword{datasets}
