% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADILSONVOL2CAP20EXER1.R
\docType{data}
\name{ADILSONVOL2CAP20EXER1}
\alias{ADILSONVOL2CAP20EXER1}
\title{ADILSONVOL2CAP20EXER1}
\format{
A data frame with 11 rows and 2 variables:
\describe{
  \item{ANO}{ANO}
  \item{RENDIMENTO MEDIO}{RENDIMENTO MEDIO EM REIAS}
}
}
\usage{
ADILSONVOL2CAP20EXER1
}
\description{
ADILSONVOL2CAP20EXER1
}
\keyword{datasets}
