% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADILSONVOL2CAP20EXER10.R
\docType{data}
\name{ADILSONVOL2CAP20EXER10}
\alias{ADILSONVOL2CAP20EXER10}
\title{ADILSONVOL2CAP20EXER10}
\format{
A data frame with 51 rows and 2 variables:
\describe{
  \item{CLIENTES}{CLIENTES}
  \item{TEMPO}{TEMPO DE ESPERA EM MINUTOS}
}
}
\usage{
ADILSONVOL2CAP20EXER10
}
\description{
TEMPO DE ESPERA
}
\keyword{datasets}
