% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADILSONVOL2CAP20EXER2.R
\docType{data}
\name{ADILSONVOL2CAP20EXER2}
\alias{ADILSONVOL2CAP20EXER2}
\title{ADILSONVOL2CAP20EXER2}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{CLUBE}{CLUBE BRASILEIRO}
  \item{PERCENTUAL DE TORCEDORES}{TORCEDORES EM PORCENTAGEM}
}
}
\usage{
ADILSONVOL2CAP20EXER2
}
\description{
CLUBE E TORCEDORES
}
\keyword{datasets}
