% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADILSONVOL2CAP20EXER4.R
\docType{data}
\name{ADILSONVOL2CAP20EXER4}
\alias{ADILSONVOL2CAP20EXER4}
\title{ADILSONVOL2CAP20EXER4}
\format{
A data frame with 5 rows and 4 variables:
\describe{
  \item{HOTEL}{HOTEL}
  \item{PRECO}{PRECO MEDIO DA DIARIA DO CASAL}
  \item{ESTRELAS}{NUMERO DE ESTRELAS}
  \item{QUARTOS}{NUMERO DE QUARTOS QUE O HOTEL POSSUI}
}
}
\usage{
ADILSONVOL2CAP20EXER4
}
\description{
HOTELARIA
}
\keyword{datasets}
