% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADILSONVOL2CAP20EXER9.R
\docType{data}
\name{ADILSONVOL2CAP20EXER9}
\alias{ADILSONVOL2CAP20EXER9}
\title{ADILSONVOL2CAP20EXER9}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{NOTAS DA TURMA X}{NOTAS QUE A TURMA TIROU}
  \item{FREQUENCIA}{FREQUNCIA}
}
}
\usage{
ADILSONVOL2CAP20EXER9
}
\description{
NOTAS DA TURMA X
}
\keyword{datasets}
