% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADILSONVOL2CAP6EXE1.R
\docType{data}
\name{ADILSONVOL2CAP6EXE1}
\alias{ADILSONVOL2CAP6EXE1}
\title{ADILSONVOL2CAP6EXE1}
\format{
A data frame with 11 rows and 2 variables:
\describe{
  \item{ANO}{ANO}
  \item{RENDIMENTO MEDIO}{RENDIMENTO MEDIO EM REAIS}
}
}
\usage{
ADILSONVOL2CAP6EXE1
}
\description{
NOTAS DA TURMA X
}
\keyword{datasets}
