% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADILSONVOL2CAP6EXE2.R
\docType{data}
\name{ADILSONVOL2CAP6EXE2}
\alias{ADILSONVOL2CAP6EXE2}
\title{ADILSONVOL2CAP6EXE2}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{CLUBE}{CLUBE BRASILEIRO}
  \item{PERCENTUAL DE TORCEDORES}{TORCEDORES EM PORCENTAGEM}
}
}
\usage{
ADILSONVOL2CAP6EXE2
}
\description{
TORCEDORES POR CLUBE
}
\keyword{datasets}
