% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALESTRIVOL2CAP7EXEM4.R
\docType{data}
\name{BALESTRIVOL2CAP7EXEM4}
\alias{BALESTRIVOL2CAP7EXEM4}
\title{BALESTRIVOL2CAP7EXEM4}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{NUMERO}{NUMERO QUE O FUNCIONARIO CALCA}
  \item{FREQUNCIA}{QUANTIDADE DE FUNCIONARIOS QUE CALCAM AQUELE NUMERO}
}
}
\usage{
BALESTRIVOL2CAP7EXEM4
}
\description{
NUMERACAO DO CALCADO DOS FUNCIONARIOS DE CERTA EMPRESA
}
\keyword{datasets}
