% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALESTRIVOL2CAP7EXEM5.R
\docType{data}
\name{BALESTRIVOL2CAP7EXEM5}
\alias{BALESTRIVOL2CAP7EXEM5}
\title{BALESTRIVOL2CAP7EXEM5}
\format{
A data frame with 46 rows and 1 variables:
\describe{
  \item{MASSA}{MASSA DO PAO FRANCES EM GRAMAS}
}
}
\usage{
BALESTRIVOL2CAP7EXEM5
}
\description{
MASSA DO PAO FRANCES
}
\keyword{datasets}
