% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALESTRIVOL2CAP7EXER12.R
\docType{data}
\name{BALESTRIVOL2CAP7EXER12}
\alias{BALESTRIVOL2CAP7EXER12}
\title{BALESTRIVOL2CAP7EXER12}
\format{
A data frame with 16 rows and 2 variables:
\describe{
  \item{DIA DA SEMANA}{DIA DA SEMANA EM QUE FOI CONSTATADO A VISITA}
  \item{FREQUENCIA}{FREQUENCIA DE VISITAS EM CADA DIA}
}
}
\usage{
BALESTRIVOL2CAP7EXER12
}
\description{
QUANTIDADE DE VISITAS DA EXPOSICAO DE ARTE
}
\keyword{datasets}
