% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALESTRIVOL2CAP7EXER13.R
\docType{data}
\name{BALESTRIVOL2CAP7EXER13}
\alias{BALESTRIVOL2CAP7EXER13}
\title{BALESTRIVOL2CAP7EXER13}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{IDADE}{IDADES DOS ALUNOS EM ANOS}
  \item{FREQUENCIA}{FREQUENCIA DOS ALUNOS COM ESSA IDADE}
}
}
\usage{
BALESTRIVOL2CAP7EXER13
}
\description{
IDADES DOS ALUNOS DO 1 ANO EM MATEMATICA EM 2016
}
\keyword{datasets}
