% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALESTRIVOL2CAP7EXER6.R
\docType{data}
\name{BALESTRIVOL2CAP7EXER6}
\alias{BALESTRIVOL2CAP7EXER6}
\title{BALESTRIVOL2CAP7EXER6}
\format{
A data frame with 5 rows and 3 variables:
\describe{
  \item{TIPO DE ENSINO}{TIPO DE ENSINO EM QUE OS ALUNOS ESTAO MATRICULADOS}
  \item{REDE PUBLICA}{ALUNOS MATRICULADOS NA REDE PUBLICA EM PORCENTAGEM}
  \item{REDE PARTICULAR}{ALUNOS MATRICULADOS NA REDE PARTICULAR EM PORCENTAGEM}
}
}
\usage{
BALESTRIVOL2CAP7EXER6
}
\description{
DISTRIBUICAO DOS ALUNOS BRASILEIROS EM 2013
}
\keyword{datasets}
