% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALESTRIVOL2CAP7EXER7.R
\docType{data}
\name{BALESTRIVOL2CAP7EXER7}
\alias{BALESTRIVOL2CAP7EXER7}
\title{BALESTRIVOL2CAP7EXER7}
\format{
A data frame with 11 rows and 2 variables:
\describe{
  \item{ANO}{ANO EM QUE FOI MEDIDO A TAXA DE DESEMPREGO}
  \item{TAXA}{TAXA DE DESEMPREGO EM PORCENTAGEM}
}
}
\usage{
BALESTRIVOL2CAP7EXER7
}
\description{
EVOLUCAO DA TAXA DE DEMPREGO
}
\keyword{datasets}
