% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras.R
\name{barras}
\alias{barras}
\title{Grafico de barras}
\usage{
barras(data = sys.frame(sys.parent()))
}
\arguments{
\item{data}{Database.}
}
\description{
Cria um grafico de barras a partir dos dados utilizados
}
