% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL1CAP1EXER19.R
\docType{data}
\name{FABIOVOL1CAP1EXER19}
\alias{FABIOVOL1CAP1EXER19}
\title{FABIOVOL1CAP1EXER19}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{DATA}{DATE}
  \item{POPULACAO}{Population of Brazil.}
}
}
\usage{
FABIOVOL1CAP1EXER19
}
\description{
POPULACAO RESIDENTE NO BRASIL
}
\keyword{datasets}
