% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL1CAP1EXER22.R
\docType{data}
\name{FABIOVOL1CAP1EXER22}
\alias{FABIOVOL1CAP1EXER22}
\title{FABIOVOL1CAP1EXER22}
\format{
A data frame with 8 rows and 2 variables:
\describe{
  \item{TEMPO}{TEMPO EM HORAS}
  \item{FREQUENCIA}{FREQUENCIA EM PERCENTUAL}

}
}
\usage{
FABIOVOL1CAP1EXER22
}
\description{
HORAS DE USO DE FONE DE OUVIDO
}
\keyword{datasets}
