% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL1CAP1EXER23.R
\docType{data}
\name{FABIOVOL1CAP1EXER23}
\alias{FABIOVOL1CAP1EXER23}
\title{FABIOVOL1CAP1EXER23}
\format{
A data frame with 37 rows and 2 variables:
\describe{
  \item{ALUNO}{ALUNO}
  \item{NOTA}{NOTA DOS ALUNOS}

}
}
\usage{
FABIOVOL1CAP1EXER23
}
\description{
NOTAS DOS ALUNOS
}
\keyword{datasets}
