% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL1CAP1EXERAUTO4.R
\docType{data}
\name{FABIOVOL1CAP1EXERAUTO4}
\alias{FABIOVOL1CAP1EXERAUTO4}
\title{FABIOVOL1CAP1EXERAUTO4}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{NOTAS DA TURMA X}{NOTA DA PROVA}
  \item{FREQUENCIA}{FREQUANCIA DE ALUNOS QUE TIRARAM ESSA NOTA}
}
}
\usage{
FABIOVOL1CAP1EXERAUTO4
}
\description{
NOTAS DA CLASSE
}
\keyword{datasets}
