% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL1CAP1EXERCOMP2.R
\docType{data}
\name{FABIOVOL1CAP1EXERCOMP2}
\alias{FABIOVOL1CAP1EXERCOMP2}
\title{FABIOVOL1CAP1EXERCOMP2}
\format{
A data frame with 6 rows and 3 variables:
\describe{
  \item{SUGESTOES}{SUGESTAO DADAS}
  \item{1A}{TURMA A EM PORCENTAGEM}
  \item{1B}{TURMA B EM PORCENTAGEM}
}
}
\usage{
FABIOVOL1CAP1EXERCOMP2
}
\description{
SUGESTAO PARA O CARDAPIO DA CANTINA
}
\keyword{datasets}
