% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL1CAP1EXERCOMP4.R
\docType{data}
\name{FABIOVOL1CAP1EXERCOMP4}
\alias{FABIOVOL1CAP1EXERCOMP4}
\title{FABIOVOL1CAP1EXERCOMP4}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{ANO}{NOTAS QUE A TURMA TIROU}
  \item{NUMERO DE ASSINATURAS}{ASSINATURA EM MILHOES}
}
}
\usage{
FABIOVOL1CAP1EXERCOMP4
}
\description{
TV POR ASSINATURA NO BRASIL
}
\keyword{datasets}
