% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXEM2.R
\docType{data}
\name{FABIOVOL3CAP3EXEM2}
\alias{FABIOVOL3CAP3EXEM2}
\title{FABIOVOL3CAP3EXEM2}
\format{
A data frame with 21 rows and 2 variables:
\describe{
  \item{ALUNO}{ALUNOS DA CLASSE DA ANA}
  \item{NOTA}{NOTA QUE OS ALUNOS TIRARAM}
}
}
\usage{
FABIOVOL3CAP3EXEM2
}
\description{
CLASSE DA ANA
}
\keyword{datasets}
