% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXEM3.R
\docType{data}
\name{FABIOVOL3CAP3EXEM3}
\alias{FABIOVOL3CAP3EXEM3}
\title{FABIOVOL3CAP3EXEM3}
\format{
A data frame with 31 rows and 2 variables:
\describe{
  \item{PESSOAS}{PESSOAS QUE TOMAM BANHO}
  \item{TEMPO}{TEMPO EM MINUTOS}
}
}
\usage{
FABIOVOL3CAP3EXEM3
}
\description{
TEMPO POR BANHO
}
\keyword{datasets}
