% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXEM7.R
\docType{data}
\name{FABIOVOL3CAP3EXEM7}
\alias{FABIOVOL3CAP3EXEM7}
\title{FABIOVOL3CAP3EXEM7}
\format{
A data frame with 14 rows and 2 variables:
\describe{
  \item{ANO}{ANO QUE FOI REALIZADO OS NUMEROS}
  \item{NUMERO DE CASOS}{QUANTIDADE DE CASOS REGISTRADOS NO ANO}
}
}
\usage{
FABIOVOL3CAP3EXEM7
}
\description{
CASOS DE DENGUE NO BRASIL
}
\keyword{datasets}
