% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXER10.R
\docType{data}
\name{FABIOVOL3CAP3EXER10}
\alias{FABIOVOL3CAP3EXER10}
\title{FABIOVOL3CAP3EXER10}
\format{
A data frame with 49 rows and 2 variables:
\describe{
  \item{SETOR}{SETOR DA EMPRESA QUE FOI MEDIDO O RUIDO}
  \item{RUIDO}{VALOR DO RUIDO}
}
}
\usage{
FABIOVOL3CAP3EXER10
}
\description{
NIVEL DE RUIDOS
}
\keyword{datasets}
