% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXER15.R
\docType{data}
\name{FABIOVOL3CAP3EXER15}
\alias{FABIOVOL3CAP3EXER15}
\title{FABIOVOL3CAP3EXER15}
\format{
A data frame with 16 rows and 2 variables:
\describe{
  \item{VEZES}{QUANTIDADE DE VEZES}
  \item{PORCENTAGEM}{PORCENTAGEM}
}
}
\usage{
FABIOVOL3CAP3EXER15
}
\description{
COM QUE FREQUENCIA
}
\keyword{datasets}
