% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXER16.R
\docType{data}
\name{FABIOVOL3CAP3EXER16}
\alias{FABIOVOL3CAP3EXER16}
\title{FABIOVOL3CAP3EXER16}
\format{
A data frame with 7 rows and 3 variables:
\describe{
  \item{REGIAO}{REGIOES DO BRASIL}
  \item{2004}{VALOR DO ANO DE 2004 EM PORCENTAGEM}
  \item{2014}{VALOR DO ANO DE 2014 EM PORCENTAGEM}
}
}
\usage{
FABIOVOL3CAP3EXER16
}
\description{
PESSOAS COM 60 ANOS OU MAIS
}
\keyword{datasets}
