% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXER19.R
\docType{data}
\name{FABIOVOL3CAP3EXER19}
\alias{FABIOVOL3CAP3EXER19}
\title{FABIOVOL3CAP3EXER19}
\format{
A data frame with 11 rows and 2 variables:
\describe{
  \item{ANO}{ANO EM QUE FORAM APURADOS OS DADOS}
  \item{VEICULOS}{QUANTIDADE DE VEICULOS NO BRASIL}
}
}
\usage{
FABIOVOL3CAP3EXER19
}
\description{
FROTA DE VEICULOS NO BRASIL
}
\keyword{datasets}
