% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXER22.R
\docType{data}
\name{FABIOVOL3CAP3EXER22}
\alias{FABIOVOL3CAP3EXER22}
\title{FABIOVOL3CAP3EXER22}
\format{
A data frame with 51 rows and 2 variables:
\describe{
  \item{APARTAMENTO}{APARTAMENTO DO PREDIO}
  \item{CONSUMO}{CONSUMO DO APARTAMENTO EM KWH}
}
}
\usage{
FABIOVOL3CAP3EXER22
}
\description{
GASTOS COM ENERGIA ELETRICA
}
\keyword{datasets}
