% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXER24.R
\docType{data}
\name{FABIOVOL3CAP3EXER24}
\alias{FABIOVOL3CAP3EXER24}
\title{FABIOVOL3CAP3EXER24}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{FACE}{FACE EM QUE O DADO SAIU}
  \item{NUMERO DE VEZES}{NUMERO DE VEZES EM QUE A FACE SAIU}
}
}
\usage{
FABIOVOL3CAP3EXER24
}
\description{
LANCAMENTO DE DADOS
}
\keyword{datasets}
