% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXER25.R
\docType{data}
\name{FABIOVOL3CAP3EXER25}
\alias{FABIOVOL3CAP3EXER25}
\title{FABIOVOL3CAP3EXER25}
\format{
A data frame with 16 rows and 2 variables:
\describe{
  \item{COR}{COR DA CAMISETA VENDIDA}
  \item{QUANTIDADE VENDIDA}{QUANTIDADE DE CAMISETAS VENDIDAS}
}
}
\usage{
FABIOVOL3CAP3EXER25
}
\description{
VENDAS DE CAMISETA
}
\keyword{datasets}
