% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXER5.R
\docType{data}
\name{FABIOVOL3CAP3EXER5}
\alias{FABIOVOL3CAP3EXER5}
\title{FABIOVOL3CAP3EXER5}
\format{
A data frame with 33 rows and 2 variables:
\describe{
  \item{ALUNO}{ALUNO DA TURMA}
  \item{CONCEITO}{CONCEITO A, B, C, D OU E}
}
}
\usage{
FABIOVOL3CAP3EXER5
}
\description{
CONCEITOS DE UMA TURMA
}
\keyword{datasets}
