% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXER6.R
\docType{data}
\name{FABIOVOL3CAP3EXER6}
\alias{FABIOVOL3CAP3EXER6}
\title{FABIOVOL3CAP3EXER6}
\format{
A data frame with 31 rows and 2 variables:
\describe{
  \item{TAXISTA}{TAXISTA}
  \item{KM}{QUANTIDADE DE KM QUE O TAXISTA PERCORREU}
}
}
\usage{
FABIOVOL3CAP3EXER6
}
\description{
KM PERCORRIDOS POR TAXISTA
}
\keyword{datasets}
