% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXER9.R
\docType{data}
\name{FABIOVOL3CAP3EXER9}
\alias{FABIOVOL3CAP3EXER9}
\title{FABIOVOL3CAP3EXER9}
\format{
A data frame with 41 rows and 2 variables:
\describe{
  \item{PESSOA}{PESSOA QUE FEZ A CONTRIBUICAO FISCAL}
  \item{VALOR}{VALOR EM REAIS}
}
}
\usage{
FABIOVOL3CAP3EXER9
}
\description{
CONTRIBUICAO FISCAL
}
\keyword{datasets}
