% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIOVOL3CAP3EXERCOMP2.R
\docType{data}
\name{FABIOVOL3CAP3EXERCOMP2}
\alias{FABIOVOL3CAP3EXERCOMP2}
\title{FABIOVOL3CAP3EXERCOMP2}
\format{
A data frame with 21 rows and 2 variables:
\describe{
  \item{POSTES}{POSTES}
  \item{ALTURA}{ALTURA DOS POSTES EM METROS}
}
}
\usage{
FABIOVOL3CAP3EXERCOMP2
}
\description{
ALTURAS DE POSTES
}
\keyword{datasets}
