% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GELSONVOL1CAP13EXER12.R
\docType{data}
\name{GELSONVOL1CAP13EXER12}
\alias{GELSONVOL1CAP13EXER12}
\title{GELSONVOL1CAP13EXER12}
\format{
A data frame with 26 rows and 2 variables:
\describe{
  \item{FUNCIONARIOS}{FUNCIONARIO}
  \item{RENDA}{RENDA EM REAIS}
}
}
\usage{
GELSONVOL1CAP13EXER12
}
\description{
RENDA MENSAL DE FUNCIONARIOS
}
\keyword{datasets}
