% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GELSONVOL1CAP13EXER13.R
\docType{data}
\name{GELSONVOL1CAP13EXER13}
\alias{GELSONVOL1CAP13EXER13}
\title{GELSONVOL1CAP13EXER13}
\format{
A data frame with 31 rows and 2 variables:
\describe{
  \item{ALUNOS}{ALUNO}
  \item{NOTA}{NOTA TIRADA NA PROVA DE INGLES}
}
}
\usage{
GELSONVOL1CAP13EXER13
}
\description{
NOTA DA PROVA DE INGLES
}
\keyword{datasets}
