% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GELSONVOL1CAP13EXER15.R
\docType{data}
\name{GELSONVOL1CAP13EXER15}
\alias{GELSONVOL1CAP13EXER15}
\title{GELSONVOL1CAP13EXER15}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{REGIAO}{REGIOES DO BRASIL}
  \item{INSCRITOS}{PESSOAS INSCRITAS POR REGIAO}
}
}
\usage{
GELSONVOL1CAP13EXER15
}
\description{
NUMEROS DE INSCRICOES NO ENEM 2013
}
\keyword{datasets}
