% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GELSONVOL1CAP13EXER16.R
\docType{data}
\name{GELSONVOL1CAP13EXER16}
\alias{GELSONVOL1CAP13EXER16}
\title{GELSONVOL1CAP13EXER16}
\format{
A data frame with 14 rows and 2 variables:
\describe{
  \item{MESES}{MESES}
  \item{VALOR}{EM BILHOES DE DOLAR}
}
}
\usage{
GELSONVOL1CAP13EXER16
}
\description{
EVOLUCAO DO SALDO DA BALANCA COMERCIAL
}
\keyword{datasets}
