% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GELSONVOL1CAP13EXER18.R
\docType{data}
\name{GELSONVOL1CAP13EXER18}
\alias{GELSONVOL1CAP13EXER18}
\title{GELSONVOL1CAP13EXER18}
\format{
A data frame with 18 rows and 2 variables:
\describe{
  \item{ANO}{ANO EM QUE ACONTECEU OS JOGOS OLIMPICOS}
  \item{MEDALHAS DE OURO}{QUANTIDADE DE MEDALHAS DE OURO CONQUISTADAS}
}
}
\usage{
GELSONVOL1CAP13EXER18
}
\description{
MEDALHAS DE OURO EM JOGOS OLIMPICOS
}
\keyword{datasets}
