% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GELSONVOL1CAP13EXER20.R
\docType{data}
\name{GELSONVOL1CAP13EXER20}
\alias{GELSONVOL1CAP13EXER20}
\title{GELSONVOL1CAP13EXER20}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{HORA}{HORA}
  \item{TEMPERATURA}{TEMPERATURA EM GRAUS CELSIUS}
}
}
\usage{
GELSONVOL1CAP13EXER20
}
\description{
TEMPERATURA DE UMA CIDADE NA SERRA GAUCHA
}
\keyword{datasets}
