% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GELSONVOL1CAP13EXER21.R
\docType{data}
\name{GELSONVOL1CAP13EXER21}
\alias{GELSONVOL1CAP13EXER21}
\title{GELSONVOL1CAP13EXER21}
\format{
A data frame with 28 rows and 2 variables:
\describe{
  \item{ESTADO}{ESTADOS BRASILEIROS}
  \item{IDMH}{IDMH INDICE DE DESENVOLVIMENTO HUMANO POR MIL}
}
}
\usage{
GELSONVOL1CAP13EXER21
}
\description{
IDMH UNIDADES DA FEDERACAO 2010
}
\keyword{datasets}
