% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADILSONVOL2CAP20EXER9.R
\docType{data}
\name{ADILSONVOL2CAP20EXER9}
\alias{ADILSONVOL2CAP20EXER9}
\title{ADILSONVOL2CAP20EXER9}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{JOGADORES}{JOGADORES DO TIME}
  \item{ALTURA}{ALTURA EM METROS}
}
}
\usage{
ADILSONVOL2CAP20EXER9
}
\description{
ALTURAS DOS JOGADORES
}
\keyword{datasets}
